I'm an idiot, part trois
It just struck me that the original goal (prime factorisation)
and my actual implementation were not the same thing. Because
I focused on the perceived "side effect" of collecting all primes
up to a point, namely sqrt(n), I was unnecessary prefiltering my
candidate factors. The original goal does not care about *all*
primes to a point, it cares only about the prime factors of
the given arg. So let's just do that.